decouple tests from source code, update configs
* refactor tests into separate files, avoid cyclic imports
* remove pytest config kluge, produce coverage file
