+++ tox.ini
@@ -32,13 +32,12 @@ deps =
     pytest-cov
     coverage_python_version
 
-commands_pre =
-    python -m pip install -U pip
-    pip install .
+#commands_pre =
+#    python -m pip install -U pip
+#    pip install .
 
 commands =
-    python -m pytest -v --cov usb_protocol --cov-branch --cov-report term-missing \
-        --keep-duplicates usb_protocol/
+    python -m pytest -v --cov usb_protocol --cov-branch --cov-report term-missing
 
 [testenv:style]
 passenv =
@@ -85,7 +84,7 @@ commands =
     python -m build .
     twine check dist/*
 
-[testenv:force]
+[testenv:dev]
 skip_install = true
 passenv =
     pythonLocation
@@ -101,10 +100,8 @@ deps =
     construct
 
 commands =
-    #python -m pip install .
-    #python -m unittest discover
-    bash -c 'export TEST_FILES=$(find usb_protocol/ -name \*.py | xargs grep -l "import unittest"); \
-        python -m unittest $TEST_FILES'
+    python -m pip install -e .
+    python -m unittest discover
 
 [testenv:mypy]
 skip_install = true
